Moved N64 flash cart code into its own directory in prep for other platforms supported on n64
